+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Only write entries that are worth mentioning to users.
 - Lib: Rename `KimiSoul.set_thinking_mode` to `KimiSoul.set_thinking`
 - UI: Better checks and notices for LLM model capabilities
 - UI: Clear the screen for `/clear` meta command
+- Tool: Support auto-downloading ripgrep on Windows
 
 ## [0.50] - 2025-11-07